config: reorganize config gen
