+++ config/config.go
@@ -1,11 +1,9 @@
 package config
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"os"
 
 	tcfg "github.com/RTradeLtd/config"
 )
@@ -87,22 +85,6 @@ func LoadConfig(configPath string) (IPFSOrchestratorConfig, error) {
 	cfg.setDefaults()
 
 	return cfg, nil
-}
-
-// GenerateConfig writes an empty orchestrator config template to given filepath
-func GenerateConfig(configPath string) error {
-	template := &IPFSOrchestratorConfig{}
-	template.setDefaults()
-	b, err := json.Marshal(template)
-	if err != nil {
-		return err
-	}
-
-	var pretty bytes.Buffer
-	if err = json.Indent(&pretty, b, "", "  "); err != nil {
-		return err
-	}
-	return ioutil.WriteFile(configPath, append(pretty.Bytes(), '\n'), os.ModePerm)
 }
 
 func (c *IPFSOrchestratorConfig) setDefaults() {
