+++ config/config_test.go
@@ -32,9 +32,3 @@ func TestLoadConfig(t *testing.T) {
 		})
 	}
 }
-
-func TestGenerateConfig(t *testing.T) {
-	if err := GenerateConfig("../config.json"); err != nil {
-		t.Error(err.Error())
-	}
-}
