+++ config/generate_test.go
@@ -1,0 +1,9 @@
+package config
+
+import "testing"
+
+func TestGenerateConfig(t *testing.T) {
+	if err := GenerateConfig("../config.json"); err != nil {
+		t.Error(err.Error())
+	}
+}
